JSON Export: only export ino for hlinks and include hlink count
Bumping the minor version of the file format to '2'.

The "ino" field is only interesting for hardlinks, so we can save space
by not exporting it for other entries.

The hlink count will be interesting later on when I implement tracking
of shared data between directories. It's currently ignored on import.
The "nlink" field makes the "hlnkc" field redundant, but let's keep
including that field anyway for backwards compatibility.